// ui-spaceage/src/app/models/partlist.ts
export class PartList {
	bomId: any | undefined;
	lot_ref_no: any | undefined;
	partNo: any | undefined;
	partDescription: any | undefined;
	version: any | undefined;
	stLoction: any | undefined;
	validity: any | undefined;
	catDescription: any | undefined;
	qtyRequired: any | undefined;
	qtyLot: any | undefined;
	primaryNo: any | undefined;
	secondaryNo: any | undefined;
	packCode: any | undefined;
	packQty: any | undefined;
	packingGroup: any | undefined;
	totalNoOfPackingGroup: any | undefined;
	mixGroup: any | undefined;
	mix: any | undefined;
	bomNo: any | undefined;
	caseMap: any | undefined;
	images: any | undefined;
	createdBy: any | undefined;
	created: any | undefined;
	modifiedBy: any | undefined;
	modified: any | undefined;
	status: any | undefined;
	barCodeNo: any | undefined;
	bomSlNo: any | undefined;
	labelStatus: any | undefined;
	colorCode: any | undefined;
	pick_label_scan: any | undefined;
	part_label_scan: any | undefined;
	enablePartLabel: any | undefined;
	totalPartScanned: any | undefined;
	netWeight: any | undefined;
	grossWeight: any | undefined;
	enableCaseReport: any | undefined;
	projectCode: any | undefined;
	pending: any | undefined;
	ack: any | undefined;
	received: any | undefined;
	packed: any | undefined;
	picByte: any | undefined;
	lot_size: any | undefined;
	deleteFlag: boolean | undefined;
}